update client service and transaction service, init Transaction and Currency services and utils, проблему маршрутизаций еще не решена

+++ transaction_service/src/main/java/com/example/transaction_service/data/model/Currency.java
@@ -20,5 +20,5 @@ public class Currency {
     private String currencyCode;
 
     @Column("currencyAmount")
-    private Long currencyAmount;
+    private Double currencyAmount;
 }
