+++ transaction_service/src/main/java/com/example/transaction_service/dto/CurrencyDto.java
@@ -15,5 +15,5 @@ public class CurrencyDto {
 
     private String currencyCode;
 
-    private Long currencyAmount;
+    private Double currencyAmount;
 }
